requirements.txt
Flask
flask-cors
openai
werkzeug
python-dotenv